requirements.txt
# Aru Archive 의존성
# Python 3.11+

# 이미지 처리
Pillow>=10.0.0

# HTTP 클라이언트 (다운로더, IPC 포워딩)
httpx>=0.27.0

# EXIF 메타데이터 (JPEG/WebP UserComment)
piexif>=1.1.3

# GUI (MVP-A 뷰어)
PyQt6>=6.6.0

# 개발/테스트
pytest>=8.0.0
